Fixed typo in ps4000.

+++ picoscope/ps4000.py
@@ -276,7 +276,7 @@ class PS4000(_PicoscopeBase):
             c_int16(self.handle), c_uint32(numPreTrigSamples),
             c_uint32(numPostTrigSamples), c_uint32(timebase),
             c_int16(oversample), byref(timeIndisposedMs),
-            c_uint32(segmentIndex), c_void_p(), c_void_p())
+            c_uint16(segmentIndex), c_void_p(), c_void_p())
         # According to the documentation, 'callback, pParameter' should work
         # instead of the last two c_void_p parameters.
         self.checkResult(m)
